+++ packages/snap/src/core/services/signer/Signer.ts
@@ -19,6 +19,7 @@ import type { Network } from '../../constants/solana';
 import type { DecompileTransactionMessageFetchingLookupTablesConfig } from '../../sdk-extensions/codecs';
 import {
   fromBytesToCompilableTransactionMessage,
+  fromUnknowBase64StringToTransaction,
   fromUnknowBase64StringToTransactionOrTransactionMessage,
 } from '../../sdk-extensions/codecs';
 import {
@@ -60,6 +61,7 @@ export class Signer {
    * @param account - The account to sign the transaction or transaction message with.
    * @param network - The network on which the transaction is being sent.
    * @param config - The configuration for the request.
+   * @param preserveMessageBytes - Whether to preserve the original message bytes.
    * @returns The signed transaction.
    * @throws If the base64 string is not a valid transaction or transaction message.
    */
@@ -68,6 +70,7 @@ export class Signer {
     account: SolanaKeyringAccount,
     network: Network,
     config?: DecompileTransactionMessageFetchingLookupTablesConfig,
+    preserveMessageBytes = false,
   ): Promise<Transaction> {
     this.#logger.log('Partially sign base64 string', {
       base64String,
@@ -75,6 +78,13 @@ export class Signer {
       network,
       config,
     });
+
+    if (preserveMessageBytes) {
+      const transaction =
+        await fromUnknowBase64StringToTransaction(base64String);
+
+      return this.#partiallySignTransaction(transaction, account);
+    }
 
     const rpc = this.#connection.getRpc(network);